[bluez] Handle also calls for unknown vcamangers
Don't skip adding a call to call list just because the call's
voicecall manager is unknown to us.

For instance, an emergency call could have a voicecall manager that
doesn't have the network registration property, so it was not added to
the known networks, but we can still handle headset requests (such as
hangup) for the call.